split the ReChart logic, colors and legend into separate files
