+++ src/components/ReChart.tsx
@@ -1,4 +1,3 @@
-import { useMemo } from "react";
 import { Box } from "@mui/material";
 import {
   PieChart,
@@ -8,42 +7,13 @@ import {
   Legend,
   ResponsiveContainer,
 } from "recharts";
-import { useTableData } from "../hooks/useTableData";
-import "@fontsource/orbitron/400.css";
+import { useReChartData } from "../hooks/useReChartData";
+import { ReChartLegend } from "./ReChartLegend";
 import { NEON_GREEN } from "../theme/ColorTheme";
-
-const mineralColors: Record<string, string> = {
-  Aetherium: "rgba(239, 115, 255, 1)",
-  Lunarium: "rgba(0, 81, 255, 1)",
-  Solenite: "rgba(0, 170, 255, 1)",
-  Nebrillium: "rgba(0, 255, 242, 1)",
-};
+import "@fontsource/orbitron/400.css";
 
 export default function RechartPieChart() {
-  const { tableData } = useTableData();
-
-  const chartData = useMemo(() => {
-    const totals = tableData.reduce((acc: Record<string, number>, row) => {
-      const mineral = row.mineral?.trim() || "Okänt";
-      const amount = Number(row.amount) || 0;
-      if (amount > 0) acc[mineral] = (acc[mineral] || 0) + amount;
-      return acc;
-    }, {});
-
-    const labels = Object.keys(totals);
-    const data = Object.values(totals);
-
-    if (labels.length === 0) {
-      labels.push("Ingen data");
-      data.push(1);
-    }
-
-    return labels.map((label, index) => ({
-      name: label,
-      value: data[index],
-      color: mineralColors[label] ?? "rgba(255,255,255,1)",
-    }));
-  }, [tableData]);
+  const chartData = useReChartData();
 
   return (
     <Box
@@ -56,26 +26,24 @@ export default function RechartPieChart() {
         justifyContent: "center",
         alignItems: "center",
         border: `1px solid ${NEON_GREEN}`,
-
         boxShadow: `
-        0 0 18px ${NEON_GREEN},    
-        0 0 8px ${NEON_GREEN},   
-        0 0 6px ${NEON_GREEN}  
-      `,
-
+          0 0 32px ${NEON_GREEN},    
+          0 0 8px ${NEON_GREEN},   
+          0 0 6px ${NEON_GREEN}  
+        `,
         py: 4.3,
       }}
     >
       <ResponsiveContainer width="100%" height="100%">
-        <PieChart width={400} height={400}>
+        <PieChart>
           <Pie
             data={chartData}
             dataKey="value"
             nameKey="name"
             cx="50%"
             cy="50%"
-            innerRadius={"40%"}
-            outerRadius={"70%"}
+            innerRadius="40%"
+            outerRadius="70%"
             paddingAngle={6}
             cornerRadius={6}
           >
@@ -97,7 +65,6 @@ export default function RechartPieChart() {
           </Pie>
 
           <Tooltip />
-
           <Legend
             layout="horizontal"
             verticalAlign="bottom"
@@ -108,48 +75,9 @@ export default function RechartPieChart() {
               fontSize: "16px",
               paddingTop: 10,
             }}
-            content={(props: any) => {
-              const { payload } = props;
-              return (
-                <div
-                  style={{
-                    display: "flex",
-                    justifyContent: "center",
-                    flexWrap: "wrap",
-                  }}
-                >
-                  {payload?.map((entry: any, index: number) => {
-                    const color =
-                      chartData.find((d) => d.name === entry.payload.name)
-                        ?.color ?? "rgba(255,255,255,1)";
-                    return (
-                      <div
-                        key={index}
-                        style={{
-                          display: "flex",
-                          alignItems: "center",
-                          margin: "4px 8px",
-                          fontFamily: "orbitron",
-                          fontSize: 16,
-                          color: "white",
-                        }}
-                      >
-                        <span
-                          style={{
-                            display: "inline-block",
-                            width: 16,
-                            height: 16,
-                            backgroundColor: color,
-                            marginRight: 6,
-                          }}
-                        />
-                        {entry.payload.name}
-                      </div>
-                    );
-                  })}
-                </div>
-              );
-            }}
+            content={(props) => (
+              <ReChartLegend {...props} chartData={chartData} />
+            )}
           />
         </PieChart>
       </ResponsiveContainer>
