+++ src/theme/MineralColors.ts
@@ -1,0 +1,6 @@
+export const mineralColors: Record<string, string> = {
+  Aetherium: "rgba(239, 115, 255, 1)",
+  Lunarium: "rgba(0, 81, 255, 1)",
+  Solenite: "rgba(0, 170, 255, 1)",
+  Nebrillium: "rgba(0, 255, 242, 1)",
+};
